requirements.txt
alembic==1.16.5
blinker==1.9.0
catboost==1.2.8
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==2.2.1
contourpy==1.3.3
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
et_xmlfile==2.0.0
Flask==3.1.2
fonttools==4.59.2
gensim==4.3.1
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
gunicorn==21.2.0
idna==3.10
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightgbm==4.6.0
Mako==1.3.10
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.8.0
mlflow==2.8.0
narwhals==2.5.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.3
packaging==23.2
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
protobuf==4.25.8
psutil==5.9.8
pyarrow==13.0.0
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
querystring-parser==1.2.4
regex==2025.9.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.10.1
seaborn==0.12.2
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
SQLAlchemy==2.0.43
sqlparse==0.5.3
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.3
xgboost==3.0.5
zipp==3.23.0
